requirements_dev.txt
-r requirements.txt
black
flake8
mypy
pre-commit
pycares==4.8.0
pytest
pytest-homeassistant-custom-component==0.13.233
pylint-pytest
pylint
pytest-aiohttp
pytest-asyncio
types-python-dateutil